Refactor: CardSection 분리
