+++ frontend/petkinFE/app/src/main/java/com/rtl/petkinfe/presentation/view/home/HomeScreen.kt
@@ -53,6 +53,10 @@ import com.rtl.petkinfe.domain.model.HealthRecord
 import com.rtl.petkinfe.domain.model.ItemType
 import com.rtl.petkinfe.domain.model.ItemTypeColors
 import com.rtl.petkinfe.domain.model.ItemTypeTitles
+import com.rtl.petkinfe.presentation.view.core.AddRecordButton
+import com.rtl.petkinfe.presentation.view.core.CardContent
+import com.rtl.petkinfe.presentation.view.core.CardHeader
+import com.rtl.petkinfe.presentation.view.core.ExpandableCard
 import com.rtl.petkinfe.presentation.view.core.IconSection
 import com.rtl.petkinfe.presentation.view.home.model.IconUIModel
 import com.rtl.petkinfe.ui.theme.PhotoIconActiveColor
@@ -153,175 +157,3 @@ fun ExpandableCardSection(
     }
 }
 
-
-@Composable
-fun ExpandableCard(
-    title: String,
-    color: Color,
-    state: CardState,
-    memo: String?,
-    onToggle: () -> Unit,
-    onPhotoUpload: () -> Unit
-) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(vertical = 12.dp)
-            .clickable { onToggle() },
-        shape = RoundedCornerShape(24.dp)
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(color)
-                .padding(12.dp)
-        ) {
-            CardHeader(title = title, isExpanded = state.isExpanded, memo =  memo, onToggle = onToggle, onAddRecord = { TODO()})
-            if (state.isExpanded) {
-                CardContent(
-                    title = title,
-                    isPhotoUploaded = state.isPhotoUploaded,
-                    memo = memo,
-                    onPhotoUpload = onPhotoUpload
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun CardHeader(
-    title: String,
-    isExpanded: Boolean,
-    memo: String?,
-    onToggle: () -> Unit,
-    onAddRecord: () -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(36.dp), // 동일한 높이로 설정
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        // 기록 여부에 따라 표시 텍스트 변경
-        Text(
-            text = if (memo == null) "기록이 없습니다" else "기록이 있습니다",
-            style = MaterialTheme.typography.bodyLarge.copy(fontSize = 16.sp),
-            fontWeight = FontWeight.Medium,
-            color = Color.Black,
-            modifier = Modifier.padding(start = 6.dp)
-        )
-        when {
-            title == "피부 질환 검사" -> {
-                // 피부 질환 검사는 열기/닫기 버튼만 표시
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    modifier = Modifier.padding(end = 6.dp)
-                    .clickable { onToggle() }
-                ) {
-                    Text(
-                        text = if (isExpanded) "닫기" else "열기",
-                        color = Color.Black,
-                        fontSize = 14.sp
-                    )
-                    Icon(
-                        imageVector = Icons.Filled.KeyboardArrowDown,
-                        contentDescription = null,
-                        tint = Color.Black
-                    )
-                }
-            }
-            memo == null -> {
-                // 기록이 없는 경우 '기록 추가' 버튼 표시
-                AddRecordButton {
-                    onAddRecord()
-                }
-            }
-            else -> {
-                // 기록이 있는 경우 '열기/닫기' 버튼 표시
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    modifier = Modifier.clickable { onToggle() }
-                ) {
-                    Text(
-                        text = if (isExpanded) "닫기" else "열기",
-                        color = Color.Black,
-                        fontSize = 14.sp
-                    )
-                    Icon(
-                        imageVector = Icons.Filled.KeyboardArrowDown,
-                        contentDescription = null,
-                        tint = Color.Black
-                    )
-                }
-            }
-        }
-    }
-}
-
-
-
-@Composable
-fun CardContent(
-    title: String,
-    isPhotoUploaded: Boolean,
-    memo: String?,
-    onPhotoUpload: () -> Unit
-) {
-    Column(modifier = Modifier.padding(16.dp)) {
-        if (title == "피부 질환 검사") {
-            TextButton(onClick = { onPhotoUpload() }) {
-                Text(
-                    text = if (!isPhotoUploaded) "사진 업로드" else "사진 업로드 완료",
-                    color = if (!isPhotoUploaded) Color.Blue else Color.Green
-                )
-            }
-        } else {
-            Text(
-                text = "📋 메모",
-                style = MaterialTheme.typography.titleMedium,
-                color = Color.Black
-            )
-            Spacer(modifier = Modifier.height(8.dp))
-            if (memo != null) {
-                Text(
-                    text = memo,
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = Color.Black
-                )
-            }
-        }
-    }
-}
-
-
-
-
-
-@Composable
-fun AddRecordButton(onClick: () -> Unit) {
-    Row(
-        modifier = Modifier
-            .height(32.dp)
-            .background(
-                color = SplashBackgroundColor, // 버튼 배경색
-                shape = RoundedCornerShape(24.dp) // 둥근 모서리
-            )
-            .border(
-                BorderStroke(0.4.dp, PhotoIconActiveColor), // 테두리
-                shape = RoundedCornerShape(24.dp)
-            )
-            .clickable(onClick = onClick) // 클릭 이벤트 추가
-            .padding(horizontal = 12.dp), // 패딩 설정
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.Center
-    ) {
-        Text(
-            text = "기록 추가",
-            fontSize = 12.sp,
-            color = Color.Black,
-            modifier = Modifier.align(Alignment.CenterVertically)
-        )
-    }
-}
